return all commit files in BlobIterator, implement getFiles in api BlobIterator now returns the files of each commit of each ref of each repo, instead of just the files on all ref head.
The schema of the files table is also changed to contain repository_id and reference_name so the filters are pushed down to the iterator and some optimizations can be done for the retrieval of just a small amount of files with certain restrictions such
as just retrieving files of a few hashes or references.
That said, this is just a temporary hack until more clever measures are put in place, because now we have two columns that will always be null and will always be dropped from the dataframe just for the sake of getting the filters pushed down.

SparkAPI has now a method `getFiles` to retrieve files given a set of repositories, references and/or hashes which takes advantage of the aforementioned optimizations until a more general approach is implemented by just using the regular dataframe of files.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>