Reuse bblfsh client instead of creating one per call
Current implementation creates one client per call. In cases where the
connections are closed at an slower pace than they are created it can
reach user limits and give errors.

This change creates a new client object the first time it is called and
returns it in the next calls.