implement MetadataSource for faster random access processing
This implements the MetadataSource, a data source that reads from in-nodes sqlite databases for improved random access performance at the expense of a small performance loss on full scans. The stored metadata is repositories, references, commits and tree_entries, since blobs are still read from siva files.

Both DefaultSource and MetadataSource share now a common test suite that will ensure both datasources are completely compatible. There is no Python counterpart to use this API as of now, but it will be done after some improvements are done regarding
the storage format and the testing of certain other parts of the code such as the JDBCIterator, etc.

There are still a few areas that could be improved such as the join between the metadata and the repositories RDDs, which is quuite heavy and expensive.
The generation of the metadata is an external process right now, that should be changed to an on-demand process, which arises some other problems: how to ensure we will get all the data cached in the right data nodes and so on.

Because of all that, this is still very early stage.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>