Parametrize repository format
Add a new configuration param to specify the format of the repositories inside the provided path. This avoids in some cases poor parallelization if there is only one folder with all the repositories inside without bucketing.

This breaks backward compatiblility, so it should be release as a new minor version (0.3.X)