[Optimizer] Throw an error if the query cannot be optimized
We cannot optimize any kind of query to push down them to the GitDatasource.

Because of that, some of the queries appears to be freezed due a lack of an optimized output.

To avoid that, now we throw an error if the query cannot be optimized to fetch repositories metadata.

Signed-off-by: Antonio Jesus Navarro Perez <antnavper@gmail.com>