get only the last commit of a reference by default This implements the ENIP-004, which makes the default behavior returning only the last commit in a reference unless specifically asked to retrieve them all. It has sort of a hack for making it work with the metadata source, since it does not make use of the CommitIterator, which has a less hacky implementation.
Also, there's a complete rewrite of the filters and how they are used in the
iterators to allow more operators than just equality. As a side effect, 'OR'
operator is no longer supported. It was supported before, but only because it
was converted to an AND (and thus being a bug).

Fixes #303.
Closes #290.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>