Fix Engine init settings
This PR fixes a simple bug I found when running the notebook on the engine docker image. Basically, the command:

```
engine = Engine(spark, "/repositories", "standard")
```

Failed with the error:

```
Py4JJavaError: An error occurred while calling o302.count.
: org.apache.hadoop.mapreduce.lib.input.InvalidInputException: Input Pattern file:/repositories/**/* matches 0 files
```
Which I  assume it has to do with the expected structure of the repositories.

Replacing that line with:

```
engine = Engine(spark, "/repositories", "siva")
```


Fixed it.